Cambios en los tipos de datos del ejercicio 14

+++ Ej Semana 2-3/14/main.cpp
@@ -5,12 +5,12 @@ x^3 + 4x        si x>0;
 x^2 - 1         si x<0;
 */
 #include <iostream>
-#include <cmath>
+#include <cmath> // std::pow
 
-int f(int x);
+double f(double x);
 
 int main(){
-    int x=0;
+    double x=0;
 
     std::cout << "Ingrese el valor de x -> ";
     std::cin >> x;
@@ -20,7 +20,7 @@ int main(){
     return 0;
 }
 
-int f(int x){
+double f(double x){
 if(x>0)         return std::pow(x,3) + (4*x); // x^3 + 4x
 else if(x<0)    return std::pow(x,2) - 1;     // x^2 - 1
